Check the abstraction facts at the widths they are installed at
Theorem-hood was established exhaustively and only to six bits, while
bv_abstraction_width defaults to sixty-four. A fact that holds below seven
bits and stops holding at thirty-two would have passed every test in the tree
and installed a clause the query does not entail -- which is the one failure
that turns a satisfiable query unsatisfiable. Several of these facts were
synthesised rather than derived, and three already carry a width restriction
for exactly that reason, so the gap was not theoretical.

The wide lit fixtures do not close it and cannot. Each asserts the negation of
one fact and expects unsat, which is what installing a clause that contradicts
the assertion produces whether or not the clause is a theorem.

So the harness now asks its questions at the widths the abstraction runs at.
Theorem-hood is sampled at eight, sixteen, thirty-two, fifty-three and
sixty-four, from the pool the premises turn on rather than uniform words; the
circuit comparison, which stopped at twelve, reaches the same widths. Both
needed the harness widened to sixty-four-bit values, which also removes the
1u << width the reference functions could not have survived at sixty-four.

And a fourth claim, which nothing anywhere made: that a fact is true of the
operation STP actually blasts. Everything else compares against four reference
functions written in the test file. This reads the result out of a model of
the circuit BBDivMod, BBMult and the prefix adder produce, so the fact meets
STP's own arithmetic -- including the totalisations for a zero divisor, which
the reference and the blaster have disagreed about before -- and the reference
functions are checked against it on the way past.

Checked by breaking the shift saturation in a way that is invisible below
eight bits: the exhaustive pass still passes, and all three of the widened
tests fail.

The six division fixtures whose query is tractable without abstraction gain
that leg, so the fact and STP's own divider have to agree on them. The other
nine say plainly that they establish the fact is offered and applied, not that
it is true, and where that is established instead.

+++ tests/unit-tests/BVSchemaCircuit_Test.cpp
@@ -101,8 +101,9 @@ uint64_t referenceRem(uint64_t x, uint64_t s, unsigned)
 }
 uint64_t referenceMul(uint64_t x, uint64_t s, unsigned width)
 {
-  return (uint64_t)(((unsigned __int128)x * (unsigned __int128)s) &
-                    (unsigned __int128)maskOf(width));
+  // Truncating a 64-bit wraparound product to `width` bits is the same value
+  // as truncating the exact one.
+  return (x * s) & maskOf(width);
 }
 uint64_t referenceAdd(uint64_t x, uint64_t s, unsigned width)
 {
